+++ e3nn/o3/_linear.py
@@ -100,6 +100,9 @@ class Linear(CodeGenMixin, torch.nn.Module):
         self,
         irreps_in,
         irreps_out,
+        *,
+        f_in: Optional[int] = None,
+        f_out: Optional[int] = None,
         internal_weights: Optional[bool] = None,
         shared_weights: Optional[bool] = None,
         instructions: Optional[List[Tuple[int, int]]] = None,
@@ -177,6 +180,8 @@ class Linear(CodeGenMixin, torch.nn.Module):
             self.irreps_out,
             self.instructions,
             self.biases,
+            f_in,
+            f_out,
             shared_weights=shared_weights,
             optimize_einsums=self._optimize_einsums
         )
@@ -187,7 +192,7 @@ class Linear(CodeGenMixin, torch.nn.Module):
         # == Generate weights ==
         if internal_weights and self.weight_numel > 0:
             assert self.shared_weights, "Having internal weights impose shared weights"
-            self.weight = torch.nn.Parameter(torch.randn(self.weight_numel))
+            self.weight = torch.nn.Parameter(torch.randn(*((f_in, f_out) if f_in is not None else ()), self.weight_numel))
         else:
             # For TorchScript, there always has to be some kind of defined .weight
             self.register_buffer('weight', torch.Tensor())
@@ -195,7 +200,7 @@ class Linear(CodeGenMixin, torch.nn.Module):
         # == Generate biases ==
         if internal_weights and self.bias_numel > 0:
             assert self.shared_weights, "Having internal weights impose shared weights"
-            self.bias = torch.nn.Parameter(torch.zeros(self.bias_numel))  # see appendix C.1 and Eq.5 of https://arxiv.org/pdf/2011.14522.pdf
+            self.bias = torch.nn.Parameter(torch.zeros(*((f_out,) if f_out is not None else ()), self.bias_numel))  # see appendix C.1 and Eq.5 of https://arxiv.org/pdf/2011.14522.pdf
         else:
             self.register_buffer('bias', torch.Tensor())
 
@@ -311,6 +316,8 @@ def _codegen_linear(
     irreps_out: o3.Irreps,
     instructions: List[Instruction],
     biases: List[bool],
+    f_in: Optional[int] = None,
+    f_out: Optional[int] = None,
     shared_weights: bool = False,
     optimize_einsums: bool = True,
 ) -> Tuple[fx.GraphModule, int, int]:
@@ -321,12 +328,19 @@ def _codegen_linear(
     ws = fx.Proxy(graph_out.placeholder('w', torch.Tensor))
     bs = fx.Proxy(graph_out.placeholder('b', torch.Tensor))
 
-    size = x.shape[:-1]
-    outsize = size + (irreps_out.dim,)
+    if f_in is None:
+        size = x.shape[:-1]
+        outsize = size + (irreps_out.dim,)
+    else:
+        size = x.shape[:-2]
+        outsize = size + (f_out, irreps_out.dim,)
 
     bias_numel = sum(mul_ir.dim for bias, mul_ir in zip(biases, irreps_out) if bias)
     if bias_numel > 0:
-        bs = bs.reshape(-1, bias_numel)
+        if f_out is None:
+            bs = bs.reshape(-1, bias_numel)
+        else:
+            bs = bs.reshape(-1, f_out, bias_numel)
 
     # = Short-circut for nothing to do =
     # We produce no code for empty instructions
@@ -340,7 +354,10 @@ def _codegen_linear(
         # 0 is weight_numel
         return fx.GraphModule({}, graph_out, "linear_forward"), 0, 0
 
-    x = x.reshape(-1, irreps_in.dim)
+    if f_in is None:
+        x = x.reshape(-1, irreps_in.dim)
+    else:
+        x = x.reshape(-1, f_in, irreps_in.dim)
     batch_out = x.shape[0]
 
     out_bias_list = []
@@ -350,22 +367,22 @@ def _codegen_linear(
             if sum(biases) == 1:
                 b = bs
             else:
-                b = bs[:, bias_index:bias_index + mul_ir_out.dim]
+                b = bs[..., bias_index:bias_index + mul_ir_out.dim]
                 bias_index += mul_ir_out.dim
-            out_bias_list += [[b.expand(batch_out, -1)]]
+            out_bias_list += [[b.expand(batch_out, -1) if f_out is None else b.expand(batch_out, f_out, -1)]]
         else:
             out_bias_list += [[]]
 
     weight_numel = sum(prod(ins.path_shape) for ins in instructions)
     if weight_numel > 0:
-        ws = ws.reshape(-1, weight_numel)
+        ws = ws.reshape(-1, weight_numel) if f_in is None else ws.reshape(-1, f_in, f_out, weight_numel)
 
     # = extract individual input irreps =
     if len(irreps_in) == 1:
-        x_list = [x.reshape(batch_out, irreps_in[0].mul, irreps_in[0].ir.dim)]
+        x_list = [x.reshape(batch_out, *(() if f_in is None else (f_in,)), irreps_in[0].mul, irreps_in[0].ir.dim)]
     else:
         x_list = [
-            x[:, i].reshape(batch_out, mul_ir.mul, mul_ir.ir.dim)
+            x[..., i].reshape(batch_out, *(() if f_in is None else (f_in,)), mul_ir.mul, mul_ir.ir.dim)
             for i, mul_ir in zip(irreps_in.slices(), irreps_in)
         ]
 
@@ -390,37 +407,40 @@ def _codegen_linear(
             w = ws
         else:
             w = ws[
-                :,
+                ...,
                 flat_weight_index:flat_weight_index + path_nweight
             ]
         w = w.reshape(
-            (() if shared_weights else (-1,)) + ins.path_shape
+            (() if shared_weights else (-1,)) + (() if f_in is None else (f_in, f_out)) + ins.path_shape
         )
         flat_weight_index += path_nweight
 
-        ein_out = torch.einsum(f"{z}uw,zui->zwi", w, x_list[ins.i_in])
+        if f_in is None:
+            ein_out = torch.einsum(f"{z}uw,zui->zwi", w, x_list[ins.i_in])
+        else:
+            ein_out = torch.einsum(f"{z}xyuw,zxui->zywi", w, x_list[ins.i_in])
         alpha = 1.0 / math.sqrt(
-            mul_ir_in.mul * sum(
+            (f_in or 1) * mul_ir_in.mul * sum(
                 1 if other_ins.i_out == ins.i_out else 0
                 for other_ins in instructions
             )
         )
         ein_out = alpha * ein_out
 
-        out_list += [ein_out.reshape(batch_out, mul_ir_out.dim)]
+        out_list += [ein_out.reshape(batch_out, *(() if f_out is None else (f_out,)), mul_ir_out.dim)]
 
     # = Return the result =
     out = [
         _sum_tensors(
             [out for ins, out in zip(instructions, out_list) if ins.i_out == i_out] + out_bias_list[i_out],
-            shape=(batch_out, mul_ir_out.dim),
+            shape=(batch_out, *(() if f_out is None else (f_out,)), mul_ir_out.dim),
             like=x
         )
         for i_out, mul_ir_out in enumerate(irreps_out)
         if mul_ir_out.mul > 0
     ]
     if len(out) > 1:
-        out = torch.cat(out, dim=1)
+        out = torch.cat(out, dim=-1)
     else:
         out = out[0]
 
@@ -438,13 +458,16 @@ def _codegen_linear(
         # See _tensor_product/_codegen.py for notes
         batchdim = 4
         example_inputs = (
-            torch.zeros((batchdim, irreps_in.dim)),
+            torch.zeros((batchdim, *(() if f_in is None else (f_in,)), irreps_in.dim)),
             torch.zeros(
                 1 if shared_weights else batchdim,
+                f_in or 1,
+                f_out or 1,
                 weight_numel,
             ),
             torch.zeros(
                 1 if shared_weights else batchdim,
+                f_out or 1,
                 bias_numel,
             ),
         )