+++ tests/o3/linear_test.py
@@ -239,3 +239,19 @@ def test_weight_view_unshared():
         m.weight_view_for_instruction(1, weights).fill_(0.0)
     out = m(inp, weights)
     assert torch.allclose(out[:, :6], torch.zeros(1))
+
+
+def test_f():
+    m = o3.Linear(
+        "0e + 1e + 2e",
+        "0e + 2x1e + 2e",
+        f_in=44,
+        f_out=5,
+        _optimize_einsums=False
+    )
+    assert_equivariant(m, args_in=[torch.randn(10, 44, 9)])
+    m = assert_auto_jitable(m)
+    y = m(torch.randn(10, 44, 9))
+    assert m.weight_numel == 4
+    assert m.weight.numel() == 44 * 5 * 4
+    assert 0.7 < y.pow(2).mean() < 1.4